Only add the private network on VirtualBox.
The private network conflicts on a multi-tenant virtualization host
since it can only route to one VM. Additionally it assumes the
VirtualBox subnet, so it's not routable on libvirt/KVM anyway.

On libvirt/KVM the guest will get a DHCP IP which can be discovered by
the user using `vagrant ssh-config`. To enable preview in Open edX, the
user can add the appropriate line to their workstation `/etc/hosts`,
rather than on the virtualization host.

For example:

    ip=$(vagrant ssh-config | awk '$1=="HostName"{print $2}')
    echo "$ip preview.localhost" | sudo tail -a /etc/hosts